+++ controllers/AdminUserController.php
@@ -106,10 +106,6 @@ class AdminUserController extends AdminModelEditorController
             ]
         ]);
 
-        if (isset($actions['related-properties']))
-        {
-            unset($actions['related-properties']);
-        }
 
         return $actions;
     }