+++ CHANGELOG.md
@@ -3,6 +3,7 @@ CHANGELOG
 
 3.0.3.pre
 -----------------
+ * The custom properties changes
  * Translation updates
  * Fixed https://github.com/skeeks-cms/cms-admin/issues/2